avoid 64bit warnings
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19132 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/opengl/gl_renderbuffer.cpp
@@ -91,7 +91,7 @@ void* csVBORenderBuffer::RenderLock (csGLRenderBufferLockType type)
   ext->glBindBufferARB (bufferTarget, bufferId);
   lastRLock = type;
 
-  return (void*)offset; // Offset for an interleaved buffer
+  return (void*) ((intptr_t) offset); // Offset for an interleaved buffer
 }
 
 //-----------------------------------------------------------------
